Add id's and classes in index and add styling
After comparing my version to the solution, I remembered you can add an
id and a class at the same time in the index file. That is a cleaner
solution, I will leave my version as is since I got the desired result
as well as to show my learning journey.